move code to index
